+++ bitcoin/src/bip158.rs
@@ -44,12 +44,12 @@ use core::fmt::{self, Display, Formatter};
 use hashes::{siphash24, Hash};
 use internals::write_err;
 
-use crate::blockdata::block::Block;
+use crate::blockdata::block::{Block, BlockHash};
 use crate::blockdata::script::Script;
 use crate::blockdata::transaction::OutPoint;
 use crate::consensus::encode::VarInt;
 use crate::consensus::{Decodable, Encodable};
-use crate::hash_types::{BlockHash, FilterHash, FilterHeader};
+use crate::hash_types::{FilterHash, FilterHeader};
 use crate::prelude::*;
 
 /// Golomb encoding parameter as in BIP-158, see also https://gist.github.com/sipa/576d5f09c3b86c3b1b75598d799fc845
@@ -554,8 +554,8 @@ mod test {
     use serde_json::Value;
 
     use super::*;
+    use crate::blockdata::block::BlockHash;
     use crate::consensus::encode::deserialize;
-    use crate::hash_types::BlockHash;
     use crate::ScriptBuf;
 
     #[test]
