+++ bitcoin/src/consensus/encode.rs
@@ -23,9 +23,9 @@ use internals::write_err;
 use io::{Cursor, Read};
 
 use crate::bip152::{PrefilledTransaction, ShortId};
-use crate::blockdata::block;
+use crate::blockdata::block::{self, BlockHash, TxMerkleNode};
 use crate::blockdata::transaction::{Transaction, TxIn, TxOut};
-use crate::hash_types::{BlockHash, FilterHash, FilterHeader, TxMerkleNode};
+use crate::hash_types::{FilterHash, FilterHeader};
 #[cfg(feature = "std")]
 use crate::p2p::{
     address::{AddrV2Message, Address},
