+++ bitcoin/src/lib.rs
@@ -122,7 +122,7 @@ pub use crate::{
     address::{Address, AddressType},
     amount::{Amount, Denomination, SignedAmount},
     bip32::XKeyIdentifier,
-    blockdata::block::{self, Block},
+    blockdata::block::{self, Block, BlockHash, TxMerkleNode, WitnessMerkleNode, WitnessCommitment},
     blockdata::constants,
     blockdata::fee_rate::FeeRate,
     blockdata::locktime::{self, absolute, relative},
@@ -137,7 +137,7 @@ pub use crate::{
     crypto::ecdsa,
     crypto::key::{self, PrivateKey, PubkeyHash, PublicKey, WPubkeyHash, XOnlyPublicKey},
     crypto::sighash::{self, LegacySighash, SegwitV0Sighash, TapSighash, TapSighashTag},
-    hash_types::{BlockHash, FilterHash, FilterHeader, TxMerkleNode, Txid, WitnessCommitment, Wtxid},
+    hash_types::{FilterHash, FilterHeader, Txid, Wtxid},
     merkle_tree::MerkleBlock,
     network::Network,
     pow::{CompactTarget, Target, Work},
