+++ bitcoin/src/merkle_tree/block.rs
@@ -43,11 +43,11 @@ use core::fmt;
 use hashes::Hash;
 
 use self::MerkleBlockError::*;
-use crate::blockdata::block::{self, Block};
+use crate::blockdata::block::{self, Block, TxMerkleNode};
 use crate::blockdata::transaction::Transaction;
 use crate::blockdata::weight::Weight;
 use crate::consensus::encode::{self, Decodable, Encodable};
-use crate::hash_types::{TxMerkleNode, Txid};
+use crate::hash_types::Txid;
 use crate::prelude::*;
 
 /// Data structure that represents a block header paired to a partial merkle tree.
@@ -540,8 +540,6 @@ mod tests {
 
     use super::*;
     use crate::consensus::encode::{deserialize, serialize};
-    #[cfg(feature = "rand-std")]
-    use crate::hash_types::TxMerkleNode;
     use crate::{Block, Txid};
 
     #[cfg(feature = "rand-std")]
